ovn-controller: Fix incremental processing of Port_Binding deletes.
If a Port_Binding is deleted from the Southbound DB and the
corresponding OVS interface is also deleted from the OVS DB, and
if both updates are received and processed by ovn-controller in
the same iteration, ovn-controller should process port_binding
delete operations first.

This commit also adds three new unixctl debug commands for
ovn-controller:
- debug/pause: pause ovn-controller processing, except unixctl commands.
- debug/resume: resume ovn-controller processing.
- debug/status: return the status of the ovn-controller processing.

These new commands are needed by the test for this scenario as without
them we have no way of ensuring predictable results. Users should not
use these commands in production. This is also why the commands are not
documented.

CC: Numan Siddique <numans@ovn.org>
Fixes: 6b0f01116bab ("ovn-controller: Handle runtime data changes in flow output engine")
Reported-by: Tim Rozet <trozet@redhat.com>
Reported-at: https://bugzilla.redhat.com/1871961
Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>