ovn: integrate bfd for static routes
Introduce the bfd reference in logical_router_static_router table
in order to check if the next-hop is properly running using the BFD
protocol. The CMS is supposed to populate bfd column with the proper
reference otherwise the BFD status is set to admin_down.
Add BFD tests in system-ovn.at.