test: fix race in ARP replies for SNAT external ips
There is a small race where gw-router-port can be running on both
hvs and both of them can reply to the arp request ending up with
two arp replies in the pcap file. This is transitory not significant
in a real deployment that can trigger a fail here. Let's use
greq -q instead.