lib: inc-proc-eng: add the capability to run partial recompute
Introduce the capability to run a recompute just on updated nodes in case
of an incremental processing abort (e.g. if the controller is not able to
write on the sb db and the change can't be processed incrementally).
This will avoid wasting CPU time on nodes that are not changed during
last run.