controller: send GARPs for LSPs in vlan-backed network scenario
When using VLAN backed networks and OVN routers leveraging the
'ovn-chassis-mac-mappings' option, the eth.src field is replaced by the
chassis mac address in order to not expose the router mac address from
different nodes and confuse the TOR switch. However doing so the TOR
switch is not able to learn the port/mac bindings for routed E/W traffic
and it is force to always flood it. Fix this issue adding the capability
to send GARP traffic for logical switch ports if the corresponding logical
switch has the ovn-lsp-garp parameter set to true in the option column.
More into about the issue can be found here [0].

[0] https://mail.openvswitch.org/pipermail/ovs-discuss/2020-September/050678.html
Reported-at: https://bugzilla.redhat.com/show_bug.cgi?id=2087779