northd: rely on new actions for lb affinity
Rely on the following new actions in order to introduce affinity timeout
support to load-balancer sessions:
- commit_lb_aff
- chk_lb_aff

Introduce the following tables in switch and router pipelines
respectively:
- S_SWITCH_IN_LB_AFF_CHECK
- S_SWITCH_IN_LB_AFF_LEARN
- S_ROUTER_IN_LB_AFF_CHECK
- S_ROUTER_IN_LB_AFF_LEARN

In this way OVN is able to dnat connections received from the same client
to a given load-balancer to the same backend if received in the affinity
timeslot.
Please note load-balancer affinity support has scalability limitation
since we need to create a new logical flow per backend.

Acked-by: Mark Michelson <mmichels@redhat.com>
Acked-by: Dumitru Ceara <dceara@redhat.com>