system-ovn: Remove route without nexthop.
for yet unclear reasons some versions of scapy are confused by this
route. It causes them to log
"MAC address to reach destination not found. Using broadcast."
instead of using the dst mac address that is already known by linux.
It then sends some packets with a broadcast dst mac which are completly
ignored by ovn. Additionally it seems to attempt arp resolution.
This slows down the requests long enough that the test runs longer than
1 second causing multiple packets to get through and failing the test.

Signed-off-by: Felix Huettner <felix.huettner@stackit.cloud>