en-global-config: Cache vxlan mode and max tunnel key.
There are going to be additional incremental nodes that wish to know if
vxlan mode is enabled. Instead of having to traverse the chassis in each
of those nodes, we can cache the vxlan status in the global config.

We can do the same with the max tunnel key based on the vxlan setting.

Signed-off-by: Mark Michelson <mmichels@redhat.com>