Datapath_Binding: Separate type and UUID external-ids.
Prior to this commit, if you wanted to find the corresponding northbound
UUID for a southbound Logical Datapath, you could find it in one of two
places:

For logical switches, it was in external-ids:logical-switch.
For logical routers, it was in external-ids:logical-router.

With this commit, we are separating the type and UUID into separate
fields. This way, no matter the type of the datapath, you can find the
UUID. And you can find the type of the datapath without having to
potentially check multiple external-ids keys to do so.

These fields are going to be used pretty heavily by northd in upcoming
patches, so instead of making them external-ids, these are now
full-fledged columns on the southbound Datapath_Binding. The UUID of the
northbound logical datapath is in a column called "nb_uuid" and the type
of the logical datapath is stored in an enumerated column called "type".

Note that there is a seemingly-unrelated change in the check packet
length test in tests/ovn.at. This test started failing because its use
of `grep` was picking up the new "nb_uuid" column accidentally. The
change to the test uses `fetch_column` since it is more precise.

Signed-off-by: Mark Michelson <mmichels@redhat.com>