en-datapath-sync: Introduce wrapper for sbrec_datapath_binding.
Not all pointers to IDL structures are the same.

When an _insert() function is called to create a pointer to an IDL
row, the pointer returned is only valid until the point that the
transaction is committed. At that point, the IDL frees the pointer. The
OVSDB will later send an update to the IDL client. At this point the IDL
will create a more permanent pointer to the row.

The en-datapath-sync node handles this for southbound datapath bindings
by using the update_sb_dp boolean to ensure that it does not hold on to
transient pointers to southbound datapath bindings. However, other engine
nodes do not (and cannot) know that the row that they are caching is
transient. As an example, the ovn_datapath struct maintains a pointer to
an sbrec_datapath_binding. If this happens to be a transient pointer
from an insertion, then this cached pointer will become invalid and may
cause a crash.

This commit remedies the problem by introducing a wrapper structure for
sbrec_datapath_binding called ovn_datapath_binding. This is a structure
whose lifetime is controlled by ovn-northd and can safely be cached by
any engine node that uses en-datapath-sync (or any of its derivative
nodes) as input.

Signed-off-by: Mark Michelson <mmichels@redhat.com>