northd: Add and delete logical routers in en-lr-nat engine node
When a router is created or destroyed add the ability to en-lr-nat
engine node to compute the new state of the lr_nat_table instead of
recalculating the lr_nat_table for the whole northbound datapabase.

I needed to change lr_nat_table_find_by_index_() function to
lr_nat_table_find_by_uuid_() because the ovn_datapath indexes are no
longer guaranteed sequential starting from zero there needed to be a way
to determine if the entry lr_nat_table->array[index] is initialized or
not. This can be determined by using the hmap of ovn_datapaths.

Reported-by: Dumitru Ceara <dceara@redhat.com>
Reported-at: https://issues.redhat.com/browse/FDP-757
Signed-off-by: Jacob Tanenbaum <jtanenba@redhat.com>